MDR_Net-pytorch-master/MDR_Net-pytorch-master/requirements.txt
numpy~=1.23.2
torch==1.10.0
torchvision~=0.13.1
Pillow~=9.2.0

sklearn~=0.0
scikit-learn~=1.1.1
opencv-python~=4.4.0.42